requirements.txt
# Requirements for Peace Speech Project Repo

# Python Version
# python==3.8.5

# Packages
numpy==1.19.1
pandas==1.1.2
scikit-learn==0.23.2
python-slugify==4.0.1
xlrd==1.2.0

matplotlib==3.3.2
seaborn==0.11.0

# Formatting
autopep8==1.5.4
pylint==2.6.0

# Testing
coverage==5.3



